add like and unlike code

+++ app/Http/Services/WebApi/CommonTraits/Likes.php
@@ -10,7 +10,7 @@ trait Likes
     public function like($service_provider_id, $type, $transactionDescription)
     {
         $transaction = new TransactionsUsers();
-        $transaction->user_id = Auth::id();
+        $transaction->user_id = (Auth::guard('customer-web')->user() ? Auth::guard('customer-web')->user()->id : Auth::id());
         $transaction->service_provider_id = $service_provider_id;
         $transaction->type = $type;
         $transaction->transaction_type = config('constants.transactions.like');
@@ -20,6 +20,6 @@ trait Likes
 
     public function unlike($service_provider_id)
     {
-        return TransactionsUsers::where([['user_id', '=', Auth::id()], ['service_provider_id', '=', $service_provider_id], ['transaction_type', '=', config('constants.transactions.like')]])->delete();
+        return TransactionsUsers::where([['user_id', '=', (Auth::guard('customer-web')->user() ? Auth::guard('customer-web')->user()->id : Auth::id())], ['service_provider_id', '=', $service_provider_id], ['transaction_type', '=', config('constants.transactions.like')]])->delete();
     }
 }
